introduce packages and refactor code
